convert architects and bandits
